Upgrading Loading and Padding
The Loading function has now been changed to either load a single filter of curves and also can load a subset of ids from the file

The Padding now only works on one filter (as it wasn't used for multiple filters anyways). Futhermore, the padded light curves can be saved so that you do not need to run it twice with QNPy.